requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pandas
numpy
requests
streamlit
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
bcrypt
matplotlib